pricebot/requirements.txt
psycopg2-binary
python-dotenv
yfinance
requests-cache
requests-ratelimiter
pyrate-limiter
